+++ src/CommandSender.php
@@ -9,6 +9,7 @@ use Lmc\Cqrs\Types\Decoder\ResponseDecoderInterface;
 use Lmc\Cqrs\Types\Exception\NoSendCommandHandlerUsedException;
 use Lmc\Cqrs\Types\Feature\ProfileableInterface;
 use Lmc\Cqrs\Types\SendCommandHandlerInterface;
+use Lmc\Cqrs\Types\Utils;
 use Lmc\Cqrs\Types\ValueObject\OnErrorCallback;
 use Lmc\Cqrs\Types\ValueObject\OnErrorInterface;
 use Lmc\Cqrs\Types\ValueObject\OnSuccessCallback;
@@ -107,11 +108,11 @@ class CommandSender implements CommandSenderInterface
                 $handler->handle(
                     $command,
                     new OnSuccessCallback(function ($response): void {
-                        $this->setIsHandled(true);
+                        $this->setIsHandled(true, $response);
                         $this->lastSuccess = $response;
                     }),
                     new OnErrorCallback(function (\Throwable $error): void {
-                        $this->setIsHandled(true);
+                        $this->setIsHandled(true, $error);
                         $this->lastError = $error;
                     }),
                 );
@@ -204,7 +205,11 @@ class CommandSender implements CommandSenderInterface
                 $profilerItem->setDuration((int) $elapsed->getDuration());
             }
 
-            $profilerItem->setHandledBy(get_class($currentHandler));
+            $profilerItem->setHandledBy(sprintf(
+                '%s<%s>',
+                Utils::getType($currentHandler),
+                $this->handledResponseType
+            ));
             $profilerItem->setDecodedBy($this->lastUsedDecoders[$currentProfilerKey->toString()] ?? []);
 
             if ($this->lastSuccess) {